Implement peers table upsert on peer announce
Rework database schema to make (peer_id, info_hash) the primary key. It
turns out that ip_port will change in a variety of circumstances, and so
is something we need to keep updating on conflict.

Note that to insert null bytes, you have to use a BYTEA column in
postgres. Null bytes can be common in IP addresses and possibly ports.
We will need to see if switching to BYTEA columns has performance costs;
I had assumed that fixed-length VARCHARs for the byte count would have
performance benefits, but never tested that.