Implement numwant limits on peer list return
Do not return more peers than the client asks for. To distribute peers,
return a contiguous subslice of the peers table starting at a random
index. The randomness will be low since the slice is contiguous, but
given different announce intervals, there should be enough churn for
this to be effective. Clients which request more peers may get the same
peer list, but we shouldn't incentivize that behavior, since it puts
added load on the tracker to little benefit.