Only return non-stale peers to clients
A stale peer is one that has not announced in the last interval, which
defaults to 60 minutes. (More exploration may be needed to determine if
there is a better interval; this may be too long). Logic will be
included to remove stale peers from the database at some point later.

There is an issue with string literals in INTERVAL and postgres
parameters, so the query is constructed in two stages.