Implement numwant testing
Set up a dummy test with 3 peers and numwant=1 to verify that we are not
sending too many peers. In the process of setting up this test I found
and fixed a bug in the implementation of numwant limiting.

This test relies on some opaque type assertions when
processing tracker responses. These kinds of assertions are only present
in the testing code, since the tracker itself does not need to ever
decode bencoded content.