Implement first attempt at peersToGive calculation
Simple algorithm: Count how many times a peer is in the swarm, and then
use that to determine the number of peers they receive, up to the number
they request as an upper bound. Includes unit tests with refactored maps
to easily set up multiple peers.

Thank you to Deep Roy for pairing on this code with me.