Add API endpoint, with authorization managed by Authorization header
The API endpoint is /api, and requires an Authorization header, tested
against a value set on startup by environmental variable.

A fragile element of this setup which may need changing is that if no
environmental variable is set, the config struct's authorization key is
set to "". This is a side effect of Go lacking a built-in optional type.
It is then the client's responsibility to make sure they are not
accepting empty headers.