Defer teardownTest so it runs even if tests fail during development
Properly using defers in each test seems preferrable to a blanket
TRUNCATE when setting up the test config. Note that due to the design of
the DbConnect function, the DB is connected to and tables are created in
one step, so we can't use a DROP TABLE when setting up the test config.