Algorithm #3: PeersForGoodSeeds
Distribute peers to clients as a function of how many completed torrents
they are seeding, and their percentage of announces with a ratio above 1
in the current session. Limitation to the current session is unavoidable
since PeerIDs change across session.

At present, I believe that this will balance out: Large seeders will
have a relatively small proportion of their torrents with a ratio above
1 in any given session, but they will also be seeding many torrents.
However, this logic may need to be revisited if we use a non-linear
smoothing function to determine how many peers to distribute.