Add peerids table, and add testing for infohashes and peerids tables
Add peerids table, with the peer_id in peers as a foreign key for this
table. For now, include nothing else in the peerids table, but future
additions will include statistics for ranking peers and for detecting
cheating.

Also add a basic test that the infohashes table downloaded key is
updated on a completion event.