Refactor handler to write announce to db at end of transaction
Various logic was inconvenient because the current announce was already
written to the database, but the current peer announcing cannot seed to
itself and should not count itself. Update the db at the end of handling
the announce after peers are distributed. This required factoring out
the info_hash check into its own function so it can be run earlier in
the handler, which makes sense anyway.