Keep track of max upload speed per peerid
We need a way to estimate the quality of a peer as a reward for a good
seeder. This is not the same as the quality of a peer as a good seeder:
For example, part of being a good seeder is seeding a lot of torrents,
but that might not be a proxy for being a good peer to receive, since
such a seeder may be congested.

Instead, the best proxy for a good peer to receive is upload speed. This
commit implements only the most basic such algorithm: It keeps track of
the max upload speed seen so far. However, due to sample sizes and
peering inconsistencies, this will need to be adjusted later to some
sort of a rolling average.

Thank you to Nikolay Karadzhov for pairing on this code.