Implement /scrape endpoint
Return a bencoded dict with information on currently tracked info_hash's
in response to a request to /scrape.

The scrape protocol also allow queries for specific info_hash's,
recommended to reduce load on the tracker. This is currently
unimplemented, pending a clean way to dynamically generate the
appropriate SQL queries.