Increase entropy in random selection of final peer list
The old algorithm, which used a random starting index for a subsequence
of the peers slice, did not have enough entropy: If clients all announce
on the same interval, future announces will have the same row order.
Then, the old algorithm will favor peers in the middle of the table,
since randomization will only adjust the start and of the slice.

Instead, use rand.Shuffle to properly randomize the peer list each time.