Implement smoothing function for return in PeersForGoodSeeds
Use tanh to construct a smoothing function which will return a minimum
of minimumPeers peers, and a maximum of numwant peers, with peers
receiving the maximum numwant if they are seeding at least one standard
deviation above the mean seed count per client in the swarm.

Thank you to Farid Rener for pairing on this code with me.