Refactor package structure into cmd/ and internal/ structure
Divide up the Go files into separate subpackages in an internal/
subdirectory. This required rethinking and clarifying the organization
of the code to avoid circular dependencies. Testing utils are now in a
dedicated testutils subpackage.

There are two outstanding issues:

First, this refactor revealed that my testing framework is not
concurrency-safe. Multiple tests in different packages try to ADD and
DROP tables from the test DB. Right now, this is being worked around by
running tests with parallel set to 1. However, I need to reconsider if I
can fix this problem, either by limiting DB access to one subpackage, or
by having each test work with a unique table, or by using a solution
like "testcontainers".

Second, most of the user-defined structs in the package ended up in the config
subpackage, because the config contains the PeerAlgorithm, and that type
in turn depends on the Announce struct. This does not really make sense,
the Announce struct belongs in the handler subpackage. I believe I can
separate these properly by switching to an interface, but I'll need to
think about how to accomplish this cleanly.

Thank you to Ian Whitlock for suggesting this kind of refactor in a code
review.