Improve logging of failures to respond to client
In ordinary cases, client errors should lead only to an appropriate
error message being sent to the client. In cases of unexpected errors or
errors responding to the client, add clearer local logging. There is
possibly still improvement to be made in this area.

Thank you to Ian Whitlock for suggesting this in code review.