Use an embed to include html templates
template.ParseFiles uses a relative path from the current directory. In
a multi-package project like this, the current directory is the
subpackage directory when testing, but the root directory when running
the whole file. I could not figure out a way to fix this, so I switched
to using an embed. This requires moving the templates into the
subpackage. Thank you to users on #go-nuts for the suggestion.

For more discussion and a solution to this issue that does not work for
me, see https://forum.golangbridge.org/t/how-to-handle-paths-for-supporting-files-in-a-package-in-go/14651