Implement /scrape endpoint with info_hash query
The BitTorrent optional spec for the /scrape endpoint includes support
for scraping only specific info_hashes as part of the query. This is in
fact recommended for clients, to reduce load on the tracker. Implement
this.

There was some trickiness involved in dynamically constructing a SQL
query which could take multiple, optional infohashes. I am not sure that
my three-stage solution to build the query with Sprintf and then prepare
it in a single step is the best way to do this.