Add ETRACKER_DISABLE_ALLOWLIST configuration option
If $ETRACKER_DISABLE_ALLOWLIST is set to "true" in the environment, the
allowlist is ignored and any info_hash used by a client in an announce
to the tracker will be allowed.

It may be worth saving state on which client added which infohashes, but
at present they are just given the name "client added".