Implement /frontendapi/infohashes backend, remove deprecated web package
The internal/web package has now been fully implemented by the new React
frontend and /frontendapi backend. Note that the code for displaying
infohashes has been simplified, at the cost of displaying them in Base64
encoding. I do not like that the infohashes /api endpoint uses hex
instead, but keeping this for now, because the /api endpoint will
eventually be deprecated as well.