Update scripts/add_infohash.py and documentation for new REST API
The Python script has been updated and simplified to work with the new
REST API, which takes a JSON object with a base64 encoded info_hash.
Testing this script revealed that it's better to restrict the API to
encrypted connections by testing for TLS in the Request object, rather
than try to detect URL.Scheme.