Change main db table names to peers, announces, infohashes
peerids renamed to peers
peers renamed to announces

This change was long over-due, since peerids are no longer used to track
clients over time. This made the old table names very confusing, and
"peers" was always a poor name for a table that tracks announces.

This is a breaking change in the DB schema, but is the first of many.
Migration will not be provided since this software is in pre-alpha.