Implement PeersForRatio, which is the new default algorithm
PeersForRatio calculates the number of peers to return as a function of
ratio and seeding percentage. Unlike previous algorithms, PeersForRatio
does not punish new clients when they enter a swarm, but only clients
that have a record of failing to seed and/or failing to have a good
ratio. At this time, total ratio above 2.0 is not further rewarded by
the algorithm. This is an intentional choice because very high ratios
can be a sign of manipulation, whether that is due to cheating or
unnecessary racing.