Implement PostTorrentFileHandler API endpoint
This allows the posting of torrent files to the tracker. Torrent files
are stored with announce key stripped and the private key added, in
preparation for redistribution to clients from the frontend. This is an
authentication-only endpoint, and so torrent files are assumed to be
well-formed. This endpoint coexists with the existing
PostInfohashHandler, so that torrents may still be added by info_hash
alone. Testing and frontend support is not yet implemented.