Implement tests for single and multifile torrent uploads
These tests involve a lot of boilerplate to set up a proper httptest
POST request with a file. It's worth more investigation if this can be
simplified in some way.

Writing these tests uncovered a bug in the multifile torrent handling;
slices and maps must be type asserted separately, rather than together.