Re-enable persistent Redis DB
Instead of relying on a non-persistent DB to flush stale announce keys,
explicitly add logic to the prune submodule to invalide the cache when
pruning stale announce keys.

As a general todo, we should implement removal of infohashes from the
allowlist, with proper cleanup of the persistent cache as well.